Avoid clashes between sed and AWS secret
The AWS secrets are strings of numbers, characters and two special characters: `/` and `+`.
This opens the possibility for a clash with the sed command we have in the Makefile.
In testing scenarios, when a new secrets gets generated for each run, it's very likely to have a `/` in there, which then breaks the sed.
Fortunately, sed is flexible and allows other command delimiters.